requirements.txt
pandas
numpy
nltk
scikit-learn
scapy
datetime
beautifulsoup4
cryptography
yara-python